Added shared folder for front & backend types
- Removed shared types from frontend & backend
- Cleaned up code
