+++ backend/package.json
@@ -27,7 +27,8 @@
     "jade": "^1.11.0",
     "joi": "^17.4.2",
     "mongoose": "^6.0.13",
-    "morgan": "~1.9.1"
+    "morgan": "~1.9.1",
+    "shared": "file:../shared"
   },
   "devDependencies": {
     "@types/cookie": "^0.4.1",
