+++ frontend/package.json
@@ -35,6 +35,7 @@
     "joi": "^17.4.2",
     "luxon": "^2.1.1",
     "pnpm": "^6.23.2",
-    "sass": "^1.44.0"
+    "sass": "^1.44.0",
+    "shared": "file:../shared"
   }
 }
