+++ frontend/src/hooks/index.ts
@@ -5,6 +5,8 @@ import User from "./user";
 export const handle = sequence(User)
 
 export const getSession: GetSession = (req) => {
+    if (req.locals.user) req.locals.user.name = `${req.locals.user.first_name} ${req.locals.user.last_name}`
+
     return {
         user: req.locals.user,
     }
