+++ frontend/src/lib/dashboard/news/Post.svelte
@@ -1,8 +1,7 @@
 <script lang="ts">
-	import type { PostData } from '$lib/interfaces/interfaces';
 	import { overlay } from '$lib/stores';
 	import { onMount } from 'svelte';
-	import {parseDate} from '$lib/utils/dateUtils';
+	import { parseDate } from '$lib/utils/dateUtils';
 
 	export let data;
 	let date = parseDate(data.date);
