+++ frontend/src/lib/types/interfaces.ts
@@ -1,0 +1,5 @@
+export interface SliderContent {
+    imgSrc: string
+    title: string
+    desc: string
+}
