+++ frontend/src/lib/types/types.ts
@@ -1,5 +1,3 @@
 export type Icons = 'alert' | 'feed' | 'home' | 'news' | 'profile' | 'support' | 'notifcations' | 'search';
 
-export type difficulty = 'easy' | 'medium' | 'advanced';
-
 export type Layout = 'column' | 'row';
