+++ shared/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "shared",
+  "version": "1.0.0",
+  "description": "Shared types & helpers for frontend & backend",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Adam Ghowiba",
+  "license": "ISC",
+  "dependencies": {
+    "mongoose": "^6.0.14"
+  }
+}
